Separate presentational components from containers
