+++ src/containers/EditBucketlist.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 
 import { getBucketlist, editBucketlist } from '../actions/action_bucketlist';
-import UpdateBucketlistForm from './UpdateBucketList';
+import UpdateBucketlistForm from '../components/UpdateBucketList';
 
 
 class EditBucketlist extends Component {
