--- src/containers/Home.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const Home = () => (
-  <div>
-    <h3>Home</h3>
-  </div>
-);
-
-export default Home;
